requirements.txt
#pytesseract
#langchain
#chromadb
#crewai
#fastapi
#uvicorn
#python-dotenv
#pillow
#pdf2image
#openai 
# 🧠 Core LLM & Agent Stack
llama-index==0.9.48.post3
llama-index-core==0.11.22
llama-index-readers-file==0.2.2
langchain==0.1.16
crewai==0.30.10
litellm==1.35.22

# 🔍 OCR & File Processing
pytesseract==0.3.10
pypdf==4.0.2

# ⚙️ Infrastructure & Backend
fastapi==0.110.2
uvicorn==0.29.0
chromadb==0.4.24
python-dotenv==1.0.1

# 📊 Streamlit & UI (Optional)
streamlit==1.29.0
pillow==10.2.0
watchdog==3.0.0

# 📦 Dependency Compatibility
tenacity==8.2.2
numpy==1.26.4
openai==1.30.1
fpdf